+++ js/src/admin/index.ts
@@ -87,6 +87,15 @@ app.initializers.add("nearata-embed-video", (app) => {
             ),
             type: "boolean",
         })
+        .registerSetting({
+            setting:
+                "nearata-embed-video.admin.settings.options.modal_enabled",
+            label: app.translator.trans(
+                "nearata-embed-video.admin.settings.options.modal_enabled"
+            ),
+            type: "boolean",
+            help: app.translator.trans("nearata-embed-video.admin.settings.options.modal_enabled_help")
+        })
         .registerPermission(
             {
                 icon: "fas fa-tag",