+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to choose which groups can view the video players.
 - `php flarum nearataEmbedVideo:purge` to remove ALL video players from posts.
 - Extended modal with ability to add, update and remove video players and related qualities.
+- Ability to disable the modal and insert the BBCode at cursor point.
 
 ## [3.1.1] - 2022-08-04